modules/location-ingester/requirements.txt
kafka-python==2.0.2
Faker==17.6.0
grpcio-tools==1.51.3